saving

+++ Source/Core/Sprite.h
@@ -9,8 +9,8 @@ class Sprite
 public:
 	Sprite(Texture* texture, int width = 0, int height = 0, glm::vec2 startPosition = glm::vec2(0.0f));
 
-	inline float GetWidth() const { return width; }
-	inline float GetHeight() const { return height; }
+	inline int GetWidth() const { return width; }
+	inline int GetHeight() const { return height; }
 	inline glm::vec2 GetStartPosition() const { return startPosition; }
 	inline float GetAspectRatio() const { return aspectRatio; }
 	inline Texture* GetTexture() const { return texture; }
